+++ src/main/java/com/djawnstj/mvcframework/annotation/Repository.java
@@ -1,6 +1,5 @@
 package com.djawnstj.mvcframework.annotation;
 
-
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
