+++ src/main/java/com/djawnstj/mvcframework/boot/MvcApplication.java
@@ -2,8 +2,6 @@ package com.djawnstj.mvcframework.boot;
 
 import com.djawnstj.mvcframework.boot.web.embbed.tomcat.TomcatWebServer;
 import com.djawnstj.mvcframework.boot.web.server.WebServer;
-import com.djawnstj.mvcframework.boot.web.servlet.user.UserRepository;
-import com.djawnstj.mvcframework.context.ApplicationContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
