--- src/main/java/com/djawnstj/mvcframework/boot/web/servlet/UserServlet.java
@@ -1,31 +1,0 @@
-package com.djawnstj.mvcframework.boot.web.servlet;
-
-import com.djawnstj.mvcframework.boot.web.servlet.user.UserRepository;
-
-import javax.servlet.RequestDispatcher;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-
-@WebServlet("/users")
-public class UserServlet extends HttpServlet {
-
-    @Override
-    public void init() throws ServletException {
-        System.out.println("UserServlet.init");
-        super.init();
-    }
-
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        System.out.println("UserServlet.doGet");
-
-        req.setAttribute("users",UserRepository.findAll());
-
-        final RequestDispatcher requestDispatcher = req.getRequestDispatcher("users.jsp");
-        requestDispatcher.forward(req, resp);
-    }
-}
