--- src/main/java/com/djawnstj/mvcframework/boot/web/servlet/user/User.java
@@ -1,28 +1,0 @@
-package com.djawnstj.mvcframework.boot.web.servlet.user;
-
-public class User {
-
-    private String id;
-    private String pw;
-
-    public User(String id, String pw) {
-        this.id = id;
-        this.pw = pw;
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getPw() {
-        return pw;
-    }
-
-    public void setPw(String pw) {
-        this.pw = pw;
-    }
-}
