--- src/main/java/com/djawnstj/mvcframework/code/PayRepository.java
@@ -1,11 +1,0 @@
-package com.djawnstj.mvcframework.code;
-
-import com.djawnstj.mvcframework.annotation.Repository;
-
-import java.util.Collection;
-import java.util.LinkedHashMap;
-import java.util.Map;
-
-//@Repository
-public class PayRepository {
-}
