--- src/main/java/com/djawnstj/mvcframework/code/PayService.java
@@ -1,10 +1,0 @@
-package com.djawnstj.mvcframework.code;
-
-public class PayService {
-
-    public PayRepository payRepository;
-
-    public PayService(final PayRepository payRepository) {
-        this.payRepository = payRepository;
-    }
-}
