--- src/main/java/com/djawnstj/mvcframework/code/User.java
@@ -1,11 +1,0 @@
-package com.djawnstj.mvcframework.code;
-
-public class User {
-    private final String id;
-    private final String pw;
-
-    public User(final String id, final String pw) {
-        this.id = id;
-        this.pw = pw;
-    }
-}
