+++ src/main/java/com/djawnstj/mvcframework/context/ApplicationContext.java
@@ -38,6 +38,8 @@ public class ApplicationContext {
 
         createBeansReferenceByConfiguration(componentClasses);
 
+        System.out.println(beanClasses);
+
         createBeans(beanClasses);
     }
