+++ src/test/java/code/PayRepository.java
@@ -1,0 +1,5 @@
+package code;
+
+//@Repository
+public class PayRepository {
+}
