+++ src/test/java/code/UserService.java
@@ -1,4 +1,4 @@
-package com.djawnstj.mvcframework.code;
+package code;
 
 import com.djawnstj.mvcframework.annotation.Autowired;
 import com.djawnstj.mvcframework.annotation.Service;
