--- src/test/java/com/djawnstj/mvcframework/boot/web/servlet/HomeServletTest.java
@@ -1,16 +1,0 @@
-package com.djawnstj.mvcframework.boot.web.servlet;
-
-import com.djawnstj.mvcframework.MvcApplicationMain;
-import org.junit.jupiter.api.Test;
-
-import javax.servlet.ServletException;
-import java.io.IOException;
-
-class HomeServletTest {
-
-    @Test
-    void HOME_SERVLET_RESPONSE() throws IOException, ClassNotFoundException {
-        // given
-        MvcApplicationMain.main(new String[0]);
-    }
-}
