+++ src/test/java/com/djawnstj/mvcframework/controller/ControllerConfig.java
@@ -1,0 +1,23 @@
+package com.djawnstj.mvcframework.controller;
+
+import com.djawnstj.mvcframework.annotation.Configuration;
+
+@Configuration
+public class ControllerConfig {
+
+//    @Bean(name = "/home")
+//    public HomeServlet homeServlet(){
+//        return new HomeServlet();
+//    }
+//
+//    @Bean(name = "/sign-up")
+//    public SignUpServlet signUpServlet(){
+//        return new SignUpServlet();
+//    }
+//
+//    @Bean(name = "/user")
+//    public UserServlet userServlet() {
+//        return new UserServlet();
+//    }
+
+}
