+++ src/test/java/com/djawnstj/mvcframework/controller/ControllerConfigTest.java
@@ -6,8 +6,6 @@ import org.junit.jupiter.api.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 class ControllerConfigTest {
 
     private static final Logger logger = LoggerFactory.getLogger(ControllerConfigTest.class);
