+++ src/test/java/study/PayServiceTest.java
@@ -3,6 +3,10 @@ package study;
 import code.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import study.code.*;
+import study.code.PayService;
+import study.code.User;
+import study.code.UserRepository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
