+++ src/test/java/study/code/PayService.java
@@ -1,0 +1,16 @@
+package study.code;
+
+public class PayService {
+
+    public final UserRepository userRepository;
+
+    public PayService(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    public void charge(long userId, long money) {
+        User user = userRepository.findById(userId);
+
+        user.getPay().add(money);
+    }
+}
