+++ src/test/java/study/code/User.java
@@ -1,0 +1,22 @@
+package study.code;
+
+import code.Pay;
+
+public class User {
+
+    private final long id;
+    private final Pay pay;
+
+    public User(long id, Pay pay) {
+        this.id = id;
+        this.pay = pay;
+    }
+
+    public long getId() {
+        return id;
+    }
+
+    public Pay getPay() {
+        return pay;
+    }
+}
