+++ src/test/java/study/code/UserRepository.java
@@ -1,0 +1,17 @@
+package study.code;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class UserRepository {
+
+    private static Map<Long, User> users = new HashMap<>();
+
+    public void save(User user) {
+        users.put(user.getId(), user);
+    }
+
+    public User findById(long userId) {
+        return users.get(userId);
+    }
+}
